Move ListVariableListener to the api package
