+++ core/optaplanner-core-impl/src/main/java/org/optaplanner/core/impl/domain/variable/index/IndexVariableListener.java
@@ -3,8 +3,8 @@ package org.optaplanner.core.impl.domain.variable.index;
 import java.util.List;
 import java.util.Objects;
 
+import org.optaplanner.core.api.domain.variable.ListVariableListener;
 import org.optaplanner.core.api.score.director.ScoreDirector;
-import org.optaplanner.core.impl.domain.variable.ListVariableListener;
 import org.optaplanner.core.impl.domain.variable.descriptor.ListVariableDescriptor;
 import org.optaplanner.core.impl.score.director.InnerScoreDirector;
