+++ core/optaplanner-core-impl/src/main/java/org/optaplanner/core/impl/domain/variable/listener/support/AbstractNotifiable.java
@@ -5,9 +5,9 @@ import java.util.Collection;
 
 import org.optaplanner.core.api.domain.solution.PlanningSolution;
 import org.optaplanner.core.api.domain.variable.AbstractVariableListener;
+import org.optaplanner.core.api.domain.variable.ListVariableListener;
 import org.optaplanner.core.api.domain.variable.VariableListener;
 import org.optaplanner.core.api.score.director.ScoreDirector;
-import org.optaplanner.core.impl.domain.variable.ListVariableListener;
 import org.optaplanner.core.impl.util.ListBasedScalingOrderedSet;
 
 /**
