+++ core/optaplanner-core-impl/src/main/java/org/optaplanner/core/impl/domain/variable/listener/support/ElementUnassignedNotification.java
@@ -1,7 +1,7 @@
 package org.optaplanner.core.impl.domain.variable.listener.support;
 
+import org.optaplanner.core.api.domain.variable.ListVariableListener;
 import org.optaplanner.core.api.score.director.ScoreDirector;
-import org.optaplanner.core.impl.domain.variable.ListVariableListener;
 
 final class ElementUnassignedNotification<Solution_> implements ListVariableNotification<Solution_> {
