+++ core/optaplanner-core-impl/src/main/java/org/optaplanner/core/impl/domain/variable/listener/support/ListVariableChangedNotification.java
@@ -1,7 +1,7 @@
 package org.optaplanner.core.impl.domain.variable.listener.support;
 
+import org.optaplanner.core.api.domain.variable.ListVariableListener;
 import org.optaplanner.core.api.score.director.ScoreDirector;
-import org.optaplanner.core.impl.domain.variable.ListVariableListener;
 
 final class ListVariableChangedNotification<Solution_> extends AbstractNotification
         implements ListVariableNotification<Solution_> {
