+++ core/optaplanner-core-impl/src/main/java/org/optaplanner/core/impl/domain/variable/listener/support/ListVariableNotification.java
@@ -1,6 +1,6 @@
 package org.optaplanner.core.impl.domain.variable.listener.support;
 
-import org.optaplanner.core.impl.domain.variable.ListVariableListener;
+import org.optaplanner.core.api.domain.variable.ListVariableListener;
 
 public interface ListVariableNotification<Solution_>
         extends Notification<Solution_, ListVariableListener<Solution_, Object, Object>> {
